better refactor
