+++ fiftyone/core/view.py
@@ -460,33 +460,6 @@ class DatasetView(foc.SampleCollection):
 
         return self.skip(start).limit(stop - start)
 
-    def _get_distributions(self, group):
-        pipeline = _DISTRIBUTION_PIPELINES[group]
-
-        # we add a sub-pipeline for each numeric as it looks like multiple
-        # buckets in a single pipeline is not supported
-        if group == "scalars":
-            _numeric_distribution_pipelines(self, pipeline)
-
-        result = list(self.aggregate(pipeline))
-
-        if group in {"labels", "scalars"}:
-            new_result = []
-            for f in result[0].values():
-                new_result += f
-            result = new_result
-
-        if group != "scalars":
-            for idx, dist in enumerate(result):
-                result[idx]["data"] = sorted(
-                    result[idx]["data"], key=lambda c: c["count"], reverse=True
-                )
-
-        return sorted(result, key=lambda d: d["name"])
-
-    def _get_facets(self):
-        return list(self.aggregate(_FACETS_PIPELINE))
-
     def _copy_with_new_stage(self, stage):
         view = copy(self)
         view._pipeline.append(stage)
@@ -498,248 +471,3 @@ class DatasetView(foc.SampleCollection):
                 return stage["$skip"]
 
         return 0
-
-
-def _numeric_bounds(view, numerics):
-    bounds_pipeline = [{"$facet": {}}]
-    for idx, (k, v) in enumerate(numerics.items()):
-        bounds_pipeline[0]["$facet"]["numeric-%d" % idx] = [
-            {
-                "$group": {
-                    "_id": k,
-                    "min": {"$min": "$%s" % k},
-                    "max": {"$max": "$%s" % k},
-                },
-            }
-        ]
-
-    return list(view.aggregate(bounds_pipeline))[0] if len(numerics) else {}
-
-
-def _numeric_distribution_pipelines(view, pipeline):
-    numerics = view._dataset.get_field_schema(ftype=fof.IntField)
-    numerics.update(view._dataset.get_field_schema(ftype=fof.FloatField))
-
-    # here we query the min and max for each numeric field
-    # unfortunately, it looks like this has to be a separate query
-    bounds = _numeric_bounds(view, numerics)
-
-    # for each numeric field, build the boundaries array with the
-    # min/max results when adding the field's sub-pipeline
-    for idx, (k, v) in enumerate(numerics.items()):
-        sub_pipeline = "numeric-%d" % idx
-        field_bounds = bounds[sub_pipeline][0]
-        mn = field_bounds["min"]
-        mx = field_bounds["max"]
-        step = (mx - mn) / 50  # magic number
-        boundaries = [mn + step * s for s in range(0, 50)]  # magic number
-
-        pipeline[0]["$facet"][sub_pipeline] = [
-            {
-                "$bucket": {
-                    "groupBy": "$%s" % k,
-                    "boundaries": boundaries,
-                    "default": "null",
-                    "output": {"count": {"$sum": 1}},
-                }
-            },
-            {
-                "$group": {
-                    "_id": k,
-                    "data": {
-                        "$push": {
-                            "key": {
-                                "$cond": [
-                                    {"$ne": ["$_id", "null"]},
-                                    {"$add": ["$_id", step / 2]},
-                                    "null",
-                                ]
-                            },
-                            "count": "$count",
-                        }
-                    },
-                }
-            },
-            {
-                "$project": {
-                    "name": k,
-                    "type": v.__class__.__name__[
-                        : -len("Field")  # grab field type from the class
-                    ].lower(),
-                    "data": "$data",
-                }
-            },
-        ]
-
-
-_DISTRIBUTION_PIPELINES = {
-    "labels": [
-        {
-            "$facet": {
-                "detections": [
-                    {"$project": {"field": {"$objectToArray": "$$ROOT"}}},
-                    {"$unwind": "$field"},
-                    {"$match": {"field.v._cls": "Detections"}},
-                    {
-                        "$project": {
-                            "field": "$field.k",
-                            "detection": "$field.v.detections",
-                        }
-                    },
-                    {"$unwind": "$detection"},
-                    {
-                        "$group": {
-                            "_id": {
-                                "field": "$field",
-                                "label": "$detection.label",
-                            },
-                            "count": {"$sum": 1},
-                        }
-                    },
-                    {
-                        "$group": {
-                            "_id": "$_id.field",
-                            "data": {
-                                "$push": {
-                                    "key": "$_id.label",
-                                    "count": "$count",
-                                }
-                            },
-                        }
-                    },
-                    {
-                        "$project": {
-                            "name": "$_id",
-                            "type": "detection",
-                            "data": "$data",
-                        }
-                    },
-                ],
-                "classifications": [
-                    {"$project": {"field": {"$objectToArray": "$$ROOT"}}},
-                    {"$unwind": "$field"},
-                    {"$match": {"field.v._cls": "Classification"}},
-                    {
-                        "$project": {
-                            "field": "$field.k",
-                            "label": "$field.v.label",
-                        }
-                    },
-                    {
-                        "$group": {
-                            "_id": {"field": "$field", "label": "$label"},
-                            "count": {"$sum": 1},
-                        }
-                    },
-                    {
-                        "$group": {
-                            "_id": "$_id.field",
-                            "data": {
-                                "$push": {
-                                    "key": "$_id.label",
-                                    "count": "$count",
-                                }
-                            },
-                        }
-                    },
-                    {
-                        "$project": {
-                            "name": "$_id",
-                            "type": "classification",
-                            "data": "$data",
-                        }
-                    },
-                ],
-            }
-        }
-    ],
-    "tags": [
-        {"$project": {"tag": "$tags"}},
-        {"$unwind": "$tag"},
-        {"$group": {"_id": "$tag", "count": {"$sum": 1}}},
-        {"$project": {"result": "$$ROOT", "field": "tags"}},
-        {
-            "$group": {
-                "_id": "$field",
-                "data": {
-                    "$push": {"key": "$result._id", "count": "$result.count"}
-                },
-            }
-        },
-        {"$project": {"name": "$_id", "type": "tag", "data": "$data"}},
-    ],
-    "scalars": [
-        {
-            "$facet": {
-                "groupings": [
-                    {"$project": {"field": {"$objectToArray": "$$ROOT"}}},
-                    {"$unwind": "$field"},
-                    {"$match": {"field.k": {"$ne": "filepath"}}},
-                    {
-                        "$project": {
-                            "field": "$field",
-                            "type": {"$type": "$field.v"},
-                        }
-                    },
-                    {"$match": {"type": {"$ne": "array"}}},
-                    {
-                        "$match": {
-                            "field.v": {"$type": "bool"}
-                        },  # @todo: support strings
-                    },
-                    {"$project": {"field": "$field.k", "label": "$field.v"}},
-                    {
-                        "$group": {
-                            "_id": {
-                                "group": "$field",
-                                "label": "$label",
-                                "type": {"$type": "$label"},
-                            },
-                            "count": {"$sum": 1},
-                        }
-                    },
-                    {
-                        "$group": {
-                            "_id": {
-                                "group": "$_id.group",
-                                "type": "$_id.type",
-                            },
-                            "data": {
-                                "$push": {
-                                    "key": "$_id.label",
-                                    "count": "$count",
-                                }
-                            },
-                        }
-                    },
-                    {
-                        "$project": {
-                            "name": "$_id.group",
-                            "type": "$_id.type",
-                            "data": "$data",
-                        }
-                    },
-                ],
-            }
-        }
-    ],
-}
-
-
-_FACETS_PIPELINE = [
-    {
-        "$facet": {
-            "tags": [
-                {"$project": {"tag": "$tags"}},
-                {
-                    "$unwind": {
-                        "path": "$tag",
-                        "preserveNullAndEmptyArrays": True,
-                    }
-                },
-                {"$group": {"_id": "$tag", "count": {"$sum": 1}}},
-                {"$sort": {"_id": 1}},
-            ],
-        }
-    }
-]
